requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
pydantic
pymupdf
faiss-cpu
sentence-transformers
openai
numpy
pandas
requests
